fix(signalk): use static mutex allocation and fix format specifier
- Replace xSemaphoreCreateMutex() with xSemaphoreCreateMutexStatic()
  for request_map_mutex_ to avoid calling FreeRTOS API during C++
  static initialization (before scheduler is running).
- Fix %d to %u for size_t in log message.

+++ src/sensesp/signalk/signalk_ws_client.cpp
@@ -304,8 +304,8 @@ void SKWSClient::on_receive_delta(uint8_t* payload, size_t length) {
   // Need to work on null-terminated strings
   constexpr size_t kMaxWsMessageSize = 4096;
   if (length > kMaxWsMessageSize) {
-    ESP_LOGW(__FILENAME__, "WebSocket message too large (%d bytes), dropping",
-             length);
+    ESP_LOGW(__FILENAME__, "WebSocket message too large (%u bytes), dropping",
+             (unsigned)length);
     return;
   }
   std::unique_ptr<char[]> buf(new char[length + 1]);
